feat(admin): add update handler to update products
chore(i18n): translate error messages to spanish in all the handlers

+++ handlers/dashboard.go
@@ -41,6 +41,7 @@ func HandleOpenDashboard(c *gin.Context,db *sql.DB){
 }
 
 func HandleAddProduct(c *gin.Context, db *sql.DB){
+
 	name := c.PostForm("product-name")
 	description := c.PostForm("product-description")
 	weight := c.PostForm("product-weight")
@@ -59,7 +60,7 @@ func HandleAddProduct(c *gin.Context, db *sql.DB){
 	VALUES (?, ?, ?, ?, ?, ?, ?)`, 
 	name, description, weight, size, price, quantity, image)
 	if err != nil{
-		c.String(http.StatusInternalServerError, "Can not create product.")
+		c.String(http.StatusInternalServerError, "No se pudo crear el producto.")
 		return
 	}
 
@@ -77,12 +78,61 @@ func HandleDeleteProduct(c *gin.Context,db *sql.DB){
 		return
 	}
 
+	name := c.PostForm("product-name")
+	description := c.PostForm("product-description")
+	weight := c.PostForm("product-weight")
+	size := c.PostForm("product-size")
+	price := c.PostForm("product-price")
+	quantity := c.PostForm("product-quantity")
+	image := c.PostForm("product-image")
+	
+	if name == "" || description == "" || weight == "" || size == "" || price == "" || quantity == "" || image == "" {
+		view.RenderFlash(c,http.StatusOK,"Todos los campos son obligatorios","info")
+		return
+	}
+
 	_, err := db.Exec(`DELETE FROM products WHERE id = ?`,productId)
 	if err != nil {
-		c.String(http.StatusInternalServerError, "Can not delete product.")
+		c.String(http.StatusInternalServerError, "No se pudo eliminar el producto.")
 	}
 
 	flash.SetMessage(c,"Producto eliminado correctamente","success")
+	c.Header("HX-Redirect","/admin/dashboard")
+	c.Status(http.StatusSeeOther)
+}
+
+func HandleUpdateProduct(c *gin.Context, db *sql.DB){
+
+	productId := c.PostForm("product-id")
+	if productId == "" {
+		view.RenderFlash(c,http.StatusOK,"No hay producto para actualizar","info")
+		return
+	}
+
+	name := c.PostForm("product-name")
+	description := c.PostForm("product-description")
+	weight := c.PostForm("product-weight")
+	size := c.PostForm("product-size")
+	price := c.PostForm("product-price")
+	quantity := c.PostForm("product-quantity")
+	image := c.PostForm("product-image")
+
+	if name == "" || description == "" || weight == "" || size == "" || price == "" || quantity == "" || image == "" {
+		view.RenderFlash(c,http.StatusOK,"Todos los campos son obligatorios","info")
+		return
+	}
+
+	_, err := db.Exec(`UPDATE products
+	SET name = ?, description = ?, weight = ?, size = ?, price = ?, quantity = ?, image_url = ? 
+	WHERE id = ?
+	`, name, description, weight, size, price, quantity, image, productId)
+
+	if err != nil {
+		c.String(http.StatusInternalServerError, "No se pudo actualizar el producto.")
+		return
+	}
+
+	flash.SetMessage(c,"Producto actualizado correctamente","success")
 	c.Header("HX-Redirect","/admin/dashboard")
 	c.Status(http.StatusSeeOther)
 }
